поправил вопросы из ревью
