+++ src/main/java/school/faang/user_service/controller/event/EventParticipationController.java
@@ -1,15 +1,34 @@
 package school.faang.user_service.controller.event;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.sevice.event.EventParticipationService;
 
 import java.util.List;
 
-public interface EventParticipationController {
-    void registerParticipant(long eventId, UserDto userDto);
+@Component
+@RequiredArgsConstructor
+public class EventParticipationController {
+    private final EventParticipationService service;
 
-    void unregisterParticipant(long eventId, UserDto userDto);
 
-    List<UserDto> getParticipant(long eventId);
+    public void registerParticipant(long eventId, UserDto userDto) {
+        service.registerParticipant(eventId, userDto.id());
+    }
 
-    Integer getParticipantsCount(long eventId);
+
+    public void unregisterParticipant(long eventId, UserDto userDto) {
+        service.unregisterParticipant(eventId, userDto.id());
+    }
+
+
+    public List<UserDto> getParticipant(long eventId) {
+        return service.getParticipant(eventId);
+    }
+
+
+    public Integer getParticipantsCount(long eventId) {
+        return service.getParticipantsCount(eventId);
+    }
 }
